requirements.txt
black==23.12.1pandas~=2.1.4Pillow==10.2.0PyQt5~=5.15.10python-docx~=1.1.0